Cleanup unneeded files.
