--- TRP_Ask_Me/lambda/custom/handlers/ConstStates.js
@@ -1,8 +1,0 @@
-const states = {
-    SEARCHMODE: "_SEARCHMODE",
-    DESCRIPTION: "_DESCRIPTION",
-    MULTIPLE_RESULTS: "_MULTIPLE_RESULTS",
-    ACTION: "_ACTION"
-};
-
-module.exports = states;
